Rename CommonAnnotatedSecurityKey to UnclassifiedLegacyCommonAnnotatedSecurityKey
